# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lander Dispatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /carriers:
    get:
      tags: [Carriers]
      operationId: listCarriers
      summary: List carriers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated carrier list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarrierListResponse'
    post:
      tags: [Carriers]
      operationId: createCarrier
      summary: Create a carrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrierInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carrier'
  /carriers/{carrierId}:
    get:
      tags: [Carriers]
      operationId: getCarrier
      summary: Get a carrier by ID
      parameters:
        - name: carrierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Carrier detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carrier'
    patch:
      tags: [Carriers]
      operationId: updateCarrier
      summary: Update a carrier
      parameters:
        - name: carrierId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrierUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carrier'
    delete:
      tags: [Carriers]
      operationId: deleteCarrier
      summary: Delete a carrier
      parameters:
        - name: carrierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
components:
...
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Dashboard
...
    Carrier:
      type: object
      required: [id, companyName, status, createdAt]
      properties:
        id:
          type: string
        companyName:
          type: string
        carrierType:
          type: [string, 'null']
        website:
          type: [string, 'null']
        usdotNumber:
          type: [string, 'null']
        mcNumber:
          type: [string, 'null']
        einNumber:
          type: [string, 'null']
        status:
          type: string
        priority:
          type: [string, 'null']
        rating:
          type: [number, 'null']
        tags:
          type: array
          items:
            type: string
        notes:
          type: [string, 'null']
        primaryContact:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        lastContact:
          type: [string, 'null']
          format: date
        companyAddress:
          type: [string, 'null']
        companyCity:
          type: [string, 'null']
        companyState:
          type: [string, 'null']
        companyZip:
          type: [string, 'null']
        fleetSize:
          type: [integer, 'null']
        trucks:
          type: [integer, 'null']
        trailers:
          type: [integer, 'null']
        drivers:
          type: [integer, 'null']
        operatingStates:
          type: array
          items:
            type: string
        serviceTypes:
          type: array
          items:
            type: string
        ratePerMile:
          type: [number, 'null']
        insuranceExpiration:
          type: [string, 'null']
          format: date
        authorityStatus:
          type: [string, 'null']
        paymentTerms:
          type: [string, 'null']
        quickPay:
          type: [boolean, 'null']
        factoringCompany:
          type: [string, 'null']
        factoringFee:
          type: [number, 'null']
        bankName:
          type: [string, 'null']
        accountHolder:
          type: [string, 'null']
        accountNumberLast4:
          type: [string, 'null']
        routingNumberLast4:
          type: [string, 'null']
        bankCity:
          type: [string, 'null']
        bankState:
          type: [string, 'null']
        zelleAccount:
          type: [string, 'null']
        cashAppAccount:
          type: [string, 'null']
        lastLoadDate:
          type: [string, 'null']
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    CarrierInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
          minLength: 1
        carrierType:
          type: string
        website:
          type: string
        usdotNumber:
          type: string
        mcNumber:
          type: string
        einNumber:
          type: string
        status:
          type: string
        priority:
          type: string
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        primaryContact:
          type: string
        phone:
          type: string
        email:
          type: string
        lastContact:
          type: string
        companyAddress:
          type: string
        companyCity:
          type: string
        companyState:
          type: string
        companyZip:
          type: string
        fleetSize:
          type: integer
        trucks:
          type: integer
        trailers:
          type: integer
        drivers:
          type: integer
        operatingStates:
          type: array
          items:
            type: string
        serviceTypes:
          type: array
          items:
            type: string
        ratePerMile:
          type: number
        insuranceExpiration:
          type: string
        authorityStatus:
          type: string
        paymentTerms:
          type: string
        quickPay:
          type: boolean
        factoringCompany:
          type: string
        factoringFee:
          type: number
        bankName:
          type: string
        accountHolder:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: string
        bankAddress:
          type: string
        bankCity:
          type: string
        bankState:
          type: string
        bankZip:
          type: string
        zelleAccount:
          type: string
        cashAppAccount:
          type: string
    CarrierUpdate:
      type: object
      properties:
        companyName:
          type: string
          minLength: 1
        carrierType:
          type: string
        website:
          type: string
        usdotNumber:
          type: string
        mcNumber:
          type: string
        einNumber:
          type: string
        status:
          type: string
        priority:
          type: string
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        primaryContact:
          type: string
        phone:
          type: string
        email:
          type: string
        lastContact:
          type: string
        companyAddress:
          type: string
        companyCity:
          type: string
        companyState:
          type: string
        companyZip:
          type: string
        fleetSize:
          type: integer
        trucks:
          type: integer
        trailers:
          type: integer
        drivers:
          type: integer
        operatingStates:
          type: array
          items:
            type: string
        serviceTypes:
          type: array
          items:
            type: string
        ratePerMile:
          type: number
        insuranceExpiration:
          type: string
        authorityStatus:
          type: string
        paymentTerms:
          type: string
        quickPay:
          type: boolean
        factoringCompany:
          type: string
        factoringFee:
          type: number
        bankName:
          type: string
        accountHolder:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: string
        bankAddress:
          type: string
        bankCity:
          type: string
        bankState:
          type: string
        bankZip:
          type: string
        zelleAccount:
          type: string
        cashAppAccount:
          type: string
    CarrierListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Carrier'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...